+++ .env
@@ -1,12 +1,10 @@
 # passed to --input-uri (see main.py)
 INPUT_URI=http://model-hosting.beng.nl/kaldi-nl-test.mp3
-# uncomment if you have downloaded s3://x-omg-daan-av/dane-asr-worker-sample-data.tar.gz
-# INPUT_URI=http://fake-hosting.beng.nl/2101608150135908031__NOS_JOURNAAL_-WON01207359.mp4
 
 # passed to --output-uri (see main.py)
 OUTPUT_URI=s3://x-omg-daan-av/assets/2101608150135908031__NOS_JOURNAAL_-WON01207359/
 
-# make sure to mount this dir into the container (see docker-compose-dane-worker.yml)
+# make sure to mount these dirs into the container (see docker-compose.yml)
 DATA_BASE_DIR=./data
 MODEL_BASE_DIR=./model
