--- data/output/testsource__testcarrier/testsource__testcarrier_provenance.json
@@ -1,48 +1,0 @@
-[
-    {
-        "activity_name": "dane-whisper-asr-worker",
-        "activity_description": "DANE worker that runs Whisper for Automatic Speech Recognition (ASR)",
-        "processing_time_ms": 43186.03515625,
-        "start_time_unix": 1721136846.9868085,
-        "parameters": {
-            "WORD_TIMESTAMPS": true,
-            "DEVICE": "cpu",
-            "VAD": true,
-            "MODEL": "tiny",
-            "BEAM_SIZE": 5,
-            "BEST_OF": 5,
-            "TEMPERATURE": "(0.0,0.2,0.4,0.6,0.8,1.0)"
-        },
-        "software_version": {
-            "dane-whisper-asr-worker": "https://github.com/beeldengeluid/dane-whisper-asr-worker/commit/"
-        },
-        "input_data": {
-            "input_file_path": "/data/input-files/testsource__testcarrier/inputfile.wav"
-        },
-        "output_data": {
-            "output_path": "/data/output-files/testsource__testcarrier",
-            "output_uri": "s3://output/folder/testsource__testcarrier//data/output-files/testsource__testcarrier/out__testsource__testcarrier.tar.gz"
-        },
-        "steps": [
-            {
-                "activity_name": "whisper_asr",
-                "activity_description": "Transcribe an audio file using Whisper",
-                "processing_time_ms": 42371.88100814819,
-                "start_time_unix": 1721136847.778497,
-                "parameters": {
-                    "WORD_TIMESTAMPS": true,
-                    "DEVICE": "cpu",
-                    "VAD": true,
-                    "MODEL": "tiny",
-                    "BEAM_SIZE": 5,
-                    "BEST_OF": 5,
-                    "TEMPERATURE": "(0.0,0.2,0.4,0.6,0.8,1.0)"
-                },
-                "software_version": "1.0.3",
-                "input_data": "/data/input-files/testsource__testcarrier/inputfile.wav",
-                "output_data": "/data/output-files/testsource__testcarrier/transcript/testsource__testcarrier.json",
-                "steps": []
-            }
-        ]
-    }
-]
