+++ docker-compose.yml
@@ -13,8 +13,8 @@ services:
         max-size: 20m
     restart: no
     # Uncomment below to use the GPU/CUDA
-    deploy:
-      resources:
-        reservations:
-          devices: 
-          - capabilities: [gpu]
+    # deploy:
+    #   resources:
+    #     reservations:
+    #       devices: 
+    #       - capabilities: [gpu]
