--- python-docker-entrypoint.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-echo "Starting DANE Whisper ASR worker"
-
-python worker.py "$@"
-
-echo "The worker has finished"
