--- scripts/intial-test-program-ids.txt
@@ -1,1 +1,0 @@
-2102003230266006831;NOS_JOURNAAL_-WON02065023
