--- scripts/transcript-example.json
@@ -1,49 +1,0 @@
-{
-    "wordTimes": [
-        1300,
-        1630,
-        1810,
-        1970,
-        2260,
-        2400,
-        2550,
-        2925,
-        3710,
-        4010,
-        4520,
-        4920,
-        5250,
-        5340,
-        5850,
-        6180,
-        6330,
-        6450,
-        6540,
-        7500,
-        7710,
-        8430,
-        8730,
-        9600,
-        9780,
-        10380,
-        10500,
-        10590,
-        11760,
-        11910,
-        12000,
-        12060,
-        12510,
-        12630,
-        12780,
-        13350,
-        13860,
-        13980,
-        14040,
-        14520
-    ],
-    "sequenceNr": 0,
-    "start": 0,
-    "fragmentId": "0001",
-    "words": "Negen uur joh boos met het nos journaal de fractieleiders rutte wilders en verhagen sturen aan op een minderheidskabinet van vvd en cda met gedoogsteun van de pvv dat is de uitkomst van hun informele gesprekken in de afgelopen week.",
-    "carrierId": "2010073021RA5-RCR1007301P"
-}
